+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## v0.1.2 (unreleased)
-- Escape dangerous control codes for bidirectional text. See also: [rustc CVE-2021-42574](https://blog.rust-lang.org/2021/11/01/cve-2021-42574.html).
+- Escape dangerous control codes for bidirectional text. See also: [CVE-2021-42574](https://blog.rust-lang.org/2021/11/01/cve-2021-42574.html).
 
 ## v0.1.1 (2021-10-14)
 - Escape unicode control characters like `U+0085 NEXT LINE` and `U+2028 LINE SEPARATOR`.